requirements.txt
alembic==1.7.4
amqp==5.0.6
anyjson==0.3.3
APScheduler==3.8.1
asn1crypto==0.24.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.2
certifi==2021.10.8
charset-normalizer==2.0.7
chromedriver==2.24.1
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==2.1.4
dataclasses==0.8
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-APScheduler==1.12.2
Flask-Login==0.5.0
Flask-Migrate==2.7.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
flower==1.0.0
gunicorn==20.0.4
humanize==3.12.0
idna==2.6
importlib-metadata==4.8.1
importlib-resources==5.4.0
IPy==1.1
itsdangerous==2.0.1
Jinja2==2.11.3
keyring==10.6.0
keyrings.alt==3.0
kombu==5.1.0
lxml==4.6.4
Mako==1.1.5
MarkupSafe==2.0.1
nmap==0.0.1
numpy==1.19.5
pandas==1.1.5
prometheus-client==0.12.0
prompt-toolkit==3.0.22
py-emails==1.2.1
pycrypto==2.6.1
pygobject==3.26.1
PyMySQL==1.0.2
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.25
redis==3.5.3
requests==2.26.0
SecretStorage==2.3.1
selenium==2.48.0
six==1.11.0
soupsieve==2.2.1
SQLAlchemy==1.3.23
tld==0.12.6
tornado==6.1
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.2
WTForms==2.3.3
xlwt==1.3.0
zipp==3.6.0